+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # 1.1.2-Beta
 - Set getBucketItemStack in ManOWar.java to correct Item
+- Update Skyris Mason and Weaponsmith to non-legacy
+- Use Processor on Skyris Mason
 
 # 1.1.1-Beta
 - Remove a bunch of unneeded stuff